refactor(user-backend-user-management): split UserService into StudentService and TutorService
The UserService in the userLib has been split into two new services and the existing UserService, StudentService and TutorService. This has been done to enable the creation and management of different user types in the system. The StudentService and TutorService provide specific functionalities for student and tutor users, respectively.
As a result other services had to be rewritten